Overhaul
I found the experiment design behind my implementation of evaluation of
vairous recommending methods erroneous. The mistake is, maybe I could
put it this way, I use part of the test data as train data while
recommending methods should not know nothing about train data. Although
this mistake might not influence the evaluation outcome when testing the
simple CF, it surely gave a leg to word2vec and other methods, which was
a piece of devastating news, because these word2vec and others have
already performed badly even with the help with test data and their
performance would surely plummet when I correct the woeful mistake.