Fix an experiment operation error
If we are going to investigate the how to enhance performance of
recommendator systems in time series perspective, maybe, we should
prepare test set so that the items in them are arranged by time. This
arrangement should be performed whether we apply doc2vec on datasets
with time consideration in mind or without, the only difference between
these two alternatives should be how the train dataset is prepared.

Previous code violated the two principles stated above(what should stay
the same and what should not). So I fixed the problems in this commit.