[Linux] Clean code. In the process, a dangerous error was fixed.
1. 'Try out surprise' and 'Try out surprise - might be obsolete - Please...' are the same. It seems SmartGit mistakenly thought 'Try out surprise' was missing and I decided to remove it from the working tree.

2. Cleaning 'Try out surprise' and so we got 'Try out SVD - cleaned'. In the process, a hidden error was found and fixed. (I thought I fixed it but I cannot be 100% sure.) The error related to some characteristics of python notebook when dealing with global variables. But let me start from the beginning.

If you try '''
>>> l1 = [1, 2]
>>> l2 = [3, 4]
>>> [(x, y) for x in filter(lambda x: x > 0, [l1[y]]) for y in l2]
[(3, 2), (3, 3)]
>>> del y
>>> [(x, y) for x in filter(lambda x: x > 0 and y > 0, [l1[y]]) for y in l2]
''' in a newly opened python interpreter, you will get an error, like this:

  Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
  NameError: name 'y' is not defined

So, I supposed we should not use outer for-loop var 'y' in the filter of inner-for loop. But I did, in 'Try out surprise', with different variable names. That was wrong and python should throw out an error. But it didn't. Maybe that I defined the counterpart of the 'y' above elsewhere as a global variable and as a result, the error was hidden from me.