[Linux] Going to do chap 3 exp 2 coefficient - inside word2vec, across variants
Note: Nothing special in this commit. So, why this commit? I was worried about that there might be some notebook code had not been saved in last commit. You know, with the auto-save feature of python notebooks, people sometimes forget to manually save code and I am not sure about the timing of auto-save.