[Linux] Chap 4 exp 4 - time overhead - SVD - done.
Time overhead collected is stored in sqlite database.

Note: When train_percent is set to 0.5, SVD prediction aborted with a MemoryError.
In search of potential solutions, I find that the implementation of SVD prediction has much to be improved. Note that I am referring to SVD prediction here rather than SVD training, the efficiency of which I rely on implementation of python surprise module.

There could be a lot which I can dwell upon...