[Linux] Minor changes.
Such as:
Graph generated earlier is too small. So, I have to add some code to make them bigger and then re-execute cells to draw graphs. As a result, some information maintained by python notebook to reflect execution order of cells is changed.

Also, missing one data point proves to be a bit troublesome -- the SVD approach aborts abnormally due to unrealistic memory requirement when training data percentage is 50% . So I increase the lower bound of training data percentage of data to be plotted to 60%.