[Linux] Observe word2vec performance when K varies.
Comparing to performance of CF, word2vec always beats CF when real K varies in range(8, 40 + 1, 4).
Note the K above is qualified with 'real'. This begs some explanation. The truth is, there is an inconsistency in the code. When 'superficial' K equals to 4, the real size of neighbors is indeed 8.

Anyway, the fact word2vec beats CF in a range of possible K values assures me. A bit.