[Linux] An very interesting bug related to probability theory.
The bug will be fixed in next commit. This commit is aimed to record it.

I want to make test size proportional to train size. But failed.
How? What's the phenomenon?
Given test_fixed_ratio 0.2, I expected that, whatever the ratio of train and full data set, the test size should be about 20% of train size.

The bug is hidden in the following code snippet.

    test = {}
    train = {}
    for userId in data:
        total_len = len(data[userId])
        if random.random() >= train_percent:
        #if 0 == random.randint(0, 2):
        #if 2 == random.randint(0, 2):
            if random.random() < (train_percent * test_fixed_ratio) / (1 - train_percent):
                test[userId] = data[userId]
        else:
            train[userId] = data[userId]

Well, later, I found that, maybe we cannot call it a bug?
Think about it.
Given 6040 users.
When train_percent is 0.9, it is not possible to make test size about 20% of test size. Because, 0.9 * 6040 + 0.9 * 6040 * 0.2 > 6040!